Raise ParserWarning not ParserError for Pandas > 1.3.5
Raise a ParserWarning to avoid the ParserError `pandas.errors.ParserError: Defining usecols with out-of-bounds indices is not allowed.` when using Pandas>1.3.5.  Now a warning is output for the problematic lines (which relate to comments in the JANAF data).

Tested on an example txt file: https://janaf.nist.gov/tables/H-066.txt